+++ crates/serial-sensors-proto/src/data_types.rs
@@ -1,10 +1,12 @@
 mod identifier;
+mod linear_ranges;
 mod scalar;
 mod vector2;
 mod vector3;
 mod vector4;
 
 pub use identifier::Identifier;
+pub use linear_ranges::LinearRanges;
 pub use scalar::ScalarData;
 pub use vector2::Vector2Data;
 pub use vector3::Vector3Data;