+++ crates/serial-sensors-proto/src/types.rs
@@ -108,4 +108,11 @@ pub struct OrientationQuaternionF32(crate::Vector4Data<f32>);
     SensorDataType, Encode, Decode, Debug, Default, Copy, Clone, Eq, PartialEq, Ord, PartialOrd,
 )]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub struct Identification(crate::Identifier);
+pub struct Identification(crate::Identifier<64>);
+
+/// Linear value range description.
+#[derive(
+    SensorDataType, Encode, Decode, Debug, Default, Copy, Clone, Eq, PartialEq, Ord, PartialOrd,
+)]
+#[cfg_attr(feature = "defmt", derive(defmt::Format))]
+pub struct LinearRanges(crate::LinearRanges);