+++ CHANGELOG.md
@@ -9,7 +9,7 @@ This project uses [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Types, where possible, now implement `Ord`, `PartialOrd` and `Hash`.
 - Added the `SensorId` type.
-- Added the `Identification` type.
+- Added the `Identification` and `LinearRanges` types for self-describing sensors.
 
 ## [0.2.3] - 2024-07-03